Add function have_email_from_user to views
