+++ moneta/src/python/db/reset_password.py
@@ -1,20 +1,11 @@
 """Forgot password views and functions."""
 from core.db import pool_manager as db
 from core import decorators, utils
+from views import forgot_password
 
 
 class ResetPassword():
     """ Class for reseting password."""
-    @staticmethod
-    def have_email_from_user(request):
-        """Get user email."""
-        if request.method == "POST":
-            try:
-                user_email = request.POST.get('email')
-            except ValueError:
-                return None
-        return user_email
-
     @staticmethod
     @decorators.retry_request()
     def find_user_by_email(user_email):
@@ -45,13 +36,12 @@ class ResetPassword():
             return None
         return new_password
 
-
 def user_exists(request):
     """Check if user does not exist."""
     if request.method == "POST":
         try:
             user_object = ResetPassword()
-            user_email = user_object.have_email_from_user(request)
+            user_email = forgot_password.have_email_from_user(request)
             have_sql = user_object.find_user_by_email(user_email)
         except ValueError:
             return None
